cập nhật toàn bộ format number giá sản phẩm

+++ src/JavaBeans/sanPhamMua.java
@@ -15,7 +15,7 @@ public class sanPhamMua extends sanPham {
 		this.soLuongMua = soLuongMua;
 	}
 	
-	public double getThanhTien() {
+	public int getThanhTien() {
 		return soLuongMua*getGiaSanPham();
 	}
